Integrated with CocoaPods
I've removed the external libraries directory and modified to project to source the existing libraries from CocoaPods. I haven't been able to find any glaring bugs, but I wouldn't be surprised if I introduced a few. In the process of adding CocoaPods, I've put the project in a workspace.

I've had to set the ARCHS build setting to ARCHS_STANDARD_32_BIT in order to successfully compile the project. I've done this because version 2.1 of Cocos2D does not support 64 bit architectures. The next version does, and I plan to try to move the project in that direction soon.